feat: add new `php8latest` to the matrix
Similar Dockerfile as `php8`, but with root image using the latest
current PHP version, which makes this a moving target.

---

```bash
diff Dockerfile.php8 Dockerfile.php8latest
```

```diff
- FROM php:8.3-apache
+ FROM php:8.4-apache
```
```diff
- COPY ./php.ini-production "$PHP_INI_DIR/php.ini"
+ COPY ./php.ini-production.php8latest "$PHP_INI_DIR/php.ini"
```

---

PHP 8.5 deprecates a few PHP INI options and bundled Opcache without
needing a separate toggle. To account for these changes, there is a
separate INI file.

These images are not meant to be used in production, but rather as a
target to test the latest PHP version easily.