ci: drop FranzDiebold/github-env-vars-action #9641
This drops the `FranzDiebold/github-env-vars-action` in favor of native
GH-Actions replacement patterns available.

ghcr.io (and other container registries for that matter) do not allow
mix-case image names, so when a the repo is forked by a user with a
mixed case username, or when reused in a repo with a mixed case repo
name, the Docker image tagging step fails.

When a lowercase username is needed it is now dynamically bash-eval'd
directly in the called line.

Notable replacements:

 - `${{ env.CI_REPOSITORY_OWNER_SLUG }}`: `$(echo '${{ github.repository }}' | tr '[:upper:]' '[:lower:]')`
 - `${{ env.CI_REPOSITORY_NAME }}`: `${{ github.event.repository.name }}`

Related: [#9639](http://team.linkorb.com/cards/9639).